7.12

+++ km/src/main/src/cn/com/sure/common/KmConstants.java
@@ -19,16 +19,16 @@ public class KmConstants {
 	
 	
 	
-	public static final Integer CODE_ID_TASK_STATUS_NOT_STARTED = 20;        //执行状态 - 尚未开始
-	public static final Integer CODE_ID_TASK_STATUS_WAITING_FOR_EXECUTING = 21;   //执行状态 - 任务已启动
+	public static final long CODE_ID_TASK_STATUS_NOT_STARTED = 20;        //执行状态 - 尚未开始
+	public static final long CODE_ID_TASK_STATUS_WAITING_FOR_EXECUTING = 21;   //执行状态 - 任务已启动
 	
-	public static final Integer CODE_ID_TASK_STATUS_EXECUTING   = 22;       //执行状态 - 正在执行
-	public static final Integer CODE_ID_TASK_STATUS_MANUAL_INTERRUPTED = 23;//执行状态 - 人工中断Manual interrupt
-	public static final Integer CODE_ID_TASK_STATUS_FINISHED = 24;          //执行状态 - 正常结束
-	public static final Integer CODE_ID_TASK_STATUS_EXCEPTION = 25;         //执行状态 - 异常结束
+	public static final long CODE_ID_TASK_STATUS_EXECUTING   = 22;       //执行状态 - 正在执行
+	public static final long CODE_ID_TASK_STATUS_MANUAL_INTERRUPTED = 23;//执行状态 - 人工中断Manual interrupt
+	public static final long CODE_ID_TASK_STATUS_FINISHED = 24;          //执行状态 - 正常结束
+	public static final long CODE_ID_TASK_STATUS_EXCEPTION = 25;         //执行状态 - 异常结束
 
-	public static final Integer CODE_ID_TASK_STATUS_MANUAL_PAUSED = 26;      //执行状态 - 人工暂停
-	public static final Integer CODE_ID_TASK_STATUS_MANUAL_RESUMED = 27;     //执行状态 - 人工继续
+	public static final long CODE_ID_TASK_STATUS_MANUAL_PAUSED = 26;      //执行状态 - 人工暂停
+	public static final long CODE_ID_TASK_STATUS_MANUAL_RESUMED = 27;     //执行状态 - 人工继续
 	
 
 }
